requirements.txt
streamlit>=1.28.0
chromadb>=0.4.0
sentence-transformers>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.0
requests>=2.32.0
numpy>=1.25.0
pandas>=2.0.0
tiktoken>=0.5.0 